requirements.txt
absl-py==0.8.1
astor==0.8.0
atari-py==0.2.6
cachetools==3.1.1
certifi==2018.11.29
cffi==1.13.2
chardet==3.0.4
cloudpickle==1.2.2
cx-Oracle==7.2.3
cycler==0.10.0
Django==2.1.5
django-common-helpers==0.9.2
django-redis==4.10.0
future==0.18.1
gast==0.2.2
gevent==1.4.0
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.24.3
gunicorn==19.9.0
gym==0.14.0
h5py==2.10.0
idna==2.8
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.10.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pyglet==1.3.2
pyparsing==2.4.5
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
scipy==1.3.1
six==1.12.0
stable-baselines==2.8.0
stem==1.7.1
sympy==1.4
TA-Lib==0.4.17



termcolor==1.1.0
urllib3==1.24.1
websocket-client==0.56.0
Werkzeug==0.16.0
wrapt==1.11.2
ccxt==1.18.1220
stochastic==0.4.0
ta==0.4.7
mpi4py
pyyaml==5.1.2
tensorforce
